% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{admm_ols}
\alias{admm_ols}
\title{OLS with ADMM}
\usage{
admm_ols(x, y, lambda, max_passes, opt, opt_tol = 1e-04)
}
\arguments{
\item{x}{predictors}

\item{y}{response in {-1, 1}}

\item{lambda}{regularization sequence}

\item{max_passes}{maximum number of passes}

\item{opt}{optimal value}

\item{opt_tol}{relative suboptimality tolerance}
}
\description{
OLS with ADMM
}
