% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pn_binom}
\alias{pn_binom}
\title{Logistic regression with proximal newton}
\usage{
pn_binom(A, b, lambda, max_passes, opt, opt_tol)
}
\arguments{
\item{lambda}{regularization sequence}

\item{max_passes}{maximum number of passes}

\item{opt}{optimal value}

\item{opt_tol}{relative suboptimality tolerance}

\item{x}{predictors}

\item{y}{response}
}
\description{
Logistic regression with proximal newton
}
