% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopesolvers-package.R
\docType{package}
\name{slopesolvers-package}
\alias{slopesolvers}
\alias{slopesolvers-package}
\title{slopesolvers: A Suite of Solvers with Associated Functions for SLOPE}
\description{
A set of solvers for SLOPE to compare performance with various
    types of input and objectives.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/slopesolvers}
  \item Report bugs at \url{https://github.com/jolars/slopesolvers/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

}
\keyword{internal}
