% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testr-package.R
\docType{package}
\name{testr-package}
\alias{testr}
\alias{testr-package}
\title{testr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Johan Larsson \email{mail@larssonjohan.com} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

}
\keyword{internal}
